osdmap: filter out nonexistent osds from map
It is possible that the crush map contains device ids that do not exist as
osds.  Filter them out of the CRUSH result.

Drop the max devices assert, as that is trivially violated by adding a new
item to the crush map beyond max_osd (via 'ceph osd crush add ...').

Furthermore, extend the osd_weight vector to match the number of CRUSH
devices so that CRUSH doesn't try and read past the end of the vector,
and mark osd_weight as mutable so we can do thi sin a const function.
This isn't really the right solution to go into master, but it's an
acceptable workaround for those who already have "corrupted" maps.

Signed-off-by: Sage Weil <sage@newdream.net>
Signed-off-by: Greg Farnum <greg@inktank.com>